requirements.txt
fastapi
uvicorn[standard]
pydantic
PyYAML